storage: Add block device in case of devicemapper storage.
Check if the rootfs of container is devicemapper block device.
If so, pass the device path as a drive to the hypervisor.
Find out the file system type. This will be needed to be passed
to hyperstart with the newcontainer command. When fstype is
provided hyperstart uses block device instead of overlay and mounts
the device with the passed fstype.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>