create: Split hyperstart init function.
The commit for checking a container's fstype worked
as the agent/hyperstart init function is executed at create as well
as start phase causing the checks for block devices to be performed
again and fstype being assigned at start phase as well.

The hyperstart init function included functionality to add hypervisor
devices along with the logic for checking for block devices,
all of which should be done only once when the pod is executing its
"create" stage.
Split the init function to only perform the required steps at creation
time. This is done by means of a new function isPodCreated().

This also means that the container fstype values are not available at
start time and should be stored in the container state to be added in a
follow-up commit.

This was also necessary to be done, since adding the state creation
logic earlier would fail for assigning indexes as createPod is now
called twice at "start" stage.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>