state: Add fields to state for storing storage related info
We are assigning the virtio disks at create and passing
the virtio block name at start. Instead of depending on an
incrementing counter to deuce this index while starting a pod,
store the block index in container state to provide more reliability.

Add fstype that we determined while inspecting the rootfs device
to the state as well. This is required since we removed the
device inspection logic in start.

Provide helper methods to store the index and fstype to filesystem.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>