sriov: Add support for physical interfaces.
For SRIOV, we need to handle physical interfaces and pass
them to qemu through vfio.
Change Endpoint to interface type and add concrete implementations
for physical and virtual interfaces.
Rename CreateNetworkEndpoint to createVirtualNetworkEndpoint to create
virtual endpoints.
If the interface found while scanning the network namespace is a
physical interface, create a physical endpoint.

Change all usages of Endpoint to check for the concrete struct types.
For cni, we create just virtual interfaces.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>