devices: Add additional parameters to devices.attach
Add parameter to devices attach interface to indicate if
attach is called during create stage. If attach is not called
during create, then the VM has been launched and device needs
to be hotplugged. Pass container as well as this may be needed to
update the container state.

Add hotplugged variable to DeviceInfo struct to store state
if device was hotplugged.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>